Added Rock Hammer (WIP)
Rock Hammer is still not 100 percent smoothed out, a few bugs are
expected. Keep in mind it will only harvest stone, cobblestone, sand,
gravel, and dirt. It will destroy any other blocks, so be careful. This
will be fixed in a future version. P.S. Accidentally derped and reverted
a commit, so had to re-revert it. (Sorry Einstein)